requirements.txt
pandas==1.1.4
plotly==4.14.3
matplotlib==3.3.4
streamlit==0.79.0
Pillow==8.2.0
